+++ src/base/Base.types.ts
@@ -27,28 +27,6 @@ interface DisplayData {
 
 export type ValueToDisplayData<Value extends string | number> = Record<Value, DisplayData>;
 
-export interface MultiValueButtonBaseProps<Value extends string | number, MultiValue extends Value | Value[] = Value> {
-    tooltip?: string;
-    value: MultiValue;
-    defaultValue?: Value;
-    values: Value[];
-    setValue: (value: MultiValue) => void;
-    valueToDisplayData: ValueToDisplayData<Value>;
-}
-
-export interface MultiValueButtonDefaultableProps<
-    Value extends string | number,
-    MultiValue extends Value | Value[] = Value,
-> extends OptionalProperty<MultiValueButtonBaseProps<Value, MultiValue>, 'value'> {
-    isDefaultable?: boolean;
-    onDefault?: () => void;
-}
-
-export type MultiValueButtonProps<Value extends string | number, MultiValue extends Value | Value[] = Value> =
-    | (MultiValueButtonBaseProps<Value, MultiValue> &
-          PropertiesNever<MultiValueButtonDefaultableProps<Value, MultiValue>>)
-    | MultiValueButtonDefaultableProps<Value, MultiValue>;
-
 export enum ScrollOffset {
     BACKWARD,
     FORWARD,