+++ src/base/components/inputs/ButtonSelectInput.tsx
@@ -10,13 +10,13 @@ import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import type { ComponentProps } from 'react';
 
-import { ButtonSelect } from '@/base/components/buttons/ButtonSelect.tsx';
+import { SelectButton } from '@/base/components/buttons/SelectButton.tsx';
 
 export const ButtonSelectInput = <Value extends string | number, MultiValue extends Value | Value[] = Value>({
     label,
     description,
     ...buttonSelectProps
-}: ComponentProps<typeof ButtonSelect<Value, MultiValue>> & { label: string; description?: string }) => (
+}: ComponentProps<typeof SelectButton<Value, MultiValue>> & { label: string; description?: string }) => (
     <Stack>
         <Typography>{label}</Typography>
         {description && (
@@ -24,6 +24,6 @@ export const ButtonSelectInput = <Value extends string | number, MultiValue exte
                 {description}
             </Typography>
         )}
-        <ButtonSelect {...buttonSelectProps} />
+        <SelectButton {...buttonSelectProps} />
     </Stack>
 );